refactor: move types to root
