+++ cyberdrop_dl/config_definitions/authentication_settings.py
@@ -1,6 +1,6 @@
 from pydantic import BaseModel, Field
 
-from cyberdrop_dl.data_structures.types import AliasModel
+from cyberdrop_dl.types import AliasModel
 
 
 class CoomerAuth(BaseModel):
