+++ cyberdrop_dl/config_definitions/global_settings.py
@@ -4,7 +4,7 @@ from datetime import timedelta
 from pydantic import BaseModel, ByteSize, Field, NonNegativeFloat, PositiveInt, field_serializer, field_validator
 from yarl import URL
 
-from cyberdrop_dl.data_structures.types import AliasModel, ByteSizeSerilized, HttpURL, NonEmptyStr
+from cyberdrop_dl.types import AliasModel, ByteSizeSerilized, HttpURL, NonEmptyStr
 from cyberdrop_dl.utils.converters import convert_to_byte_size
 from cyberdrop_dl.utils.validators import parse_duration_as_timedelta, parse_falsy_as
