+++ cyberdrop_dl/crawlers/box_dot_com.py
@@ -10,8 +10,8 @@ from pydantic import Field
 from yarl import URL
 
 from cyberdrop_dl.crawlers.crawler import Crawler, create_task_id
-from cyberdrop_dl.data_structures.types import AliasModel
 from cyberdrop_dl.exceptions import ScrapeError
+from cyberdrop_dl.types import AliasModel
 from cyberdrop_dl.utils.utilities import error_handling_wrapper
 
 if TYPE_CHECKING:
