+++ cyberdrop_dl/crawlers/xhamster.py
@@ -7,7 +7,7 @@ from pydantic import AliasPath, Field, PlainValidator
 from yarl import URL
 
 from cyberdrop_dl.crawlers.crawler import Crawler, create_task_id
-from cyberdrop_dl.data_structures.types import AliasModel
+from cyberdrop_dl.types import AliasModel
 from cyberdrop_dl.utils import javascript
 from cyberdrop_dl.utils.logger import log_debug
 from cyberdrop_dl.utils.utilities import error_handling_wrapper, get_text_between, parse_url
