+++ cyberdrop_dl/utils/apprise.py
@@ -14,7 +14,7 @@ import rich
 from pydantic import ValidationError
 from rich.text import Text
 
-from cyberdrop_dl.data_structures.types import AppriseURLModel
+from cyberdrop_dl.types import AppriseURLModel
 from cyberdrop_dl.utils import constants
 from cyberdrop_dl.utils.logger import log, log_debug, log_spacer
 from cyberdrop_dl.utils.yaml import handle_validation_error
